Move types to types workspace
