+++ frontend/src/Grid.tsx
@@ -2,28 +2,7 @@ import React, { useEffect, useCallback, useState } from "react";
 import "./Grid.css";
 import "./Crossword.css";
 
-import { WriteMode } from "./Crossword";
-
-// Define array type that only accepts a specified number of elements
-// https://stackoverflow.com/a/59906630
-type ArrayLengthMutationKeys =
-  | "splice"
-  | "push"
-  | "pop"
-  | "shift"
-  | "unshift"
-  | number;
-type ArrayItems<T extends Array<any>> = T extends Array<infer TItems>
-  ? TItems
-  : never;
-type FixedLengthArray<T extends any[]> = Pick<
-  T,
-  Exclude<keyof T, ArrayLengthMutationKeys>
-> & { [Symbol.iterator]: () => IterableIterator<ArrayItems<T>> };
-
-export type Vec2 = FixedLengthArray<[number, number]>;
-export type Vec3 = FixedLengthArray<[number, number, number]>;
-export type Vec4 = FixedLengthArray<[number, number, number, number]>;
+import { Square, LetterType, Vec2, WriteMode } from "types";
 
 interface GridProps {
   squares: {
@@ -182,11 +161,6 @@ const Grid: React.FC<GridProps> = (props) => {
 
 export default Grid;
 
-export type Square = {
-  c: Vec4; // [x, y, width, height]
-  t: 0 | 1; // fillable or not
-};
-
 export const createCoordinateGrid = (squares: Square[][]): number[][] => {
   const grid: number[][] = [];
   let idx = 0;
@@ -199,8 +173,6 @@ export const createCoordinateGrid = (squares: Square[][]): number[][] => {
   });
   return grid;
 };
-
-export type LetterType = { letter: string; row: number; column: number };
 
 export const createLetterArray = (squares: Square[][]) => {
   const arr: LetterType[] = [];
