+++ types/package.json
@@ -1,6 +1,7 @@
 {
   "name": "types",
   "version": "0.1.0",
+  "main": "dist/index.js",
   "private": true,
   "dependencies": {},
   "devDependencies": {
@@ -9,6 +10,5 @@
   },
   "scripts": {
     "start": "tsc --watch"
-  },
-  "types": "dist/*"
+  }
 }
